+++ semTools/R/auxiliary.R
@@ -1,6 +1,5 @@
 ### Terrence D. Jorgensen
-### Last updated: 10 May 2022
-### new auxiliary function does NOT create a lavaanStar-class instance
+### Last updated: 23 May 2022
 
 ##' Implement Saturated Correlates with FIML
 ##'
@@ -10,14 +9,14 @@
 ##' These functions are wrappers around the corresponding lavaan functions.
 ##' You can use them the same way you use \code{\link[lavaan]{lavaan}}, but you
 ##' \emph{must} pass your full \code{data.frame} to the \code{data} argument.
-##' Because the saturated-correlates approaches (Enders, 2008) treates exogenous
+##' Because the saturated-correlates approaches (Enders, 2008) treats exogenous
 ##' variables as random, \code{fixed.x} must be set to \code{FALSE}. Because FIML
 ##' requires continuous data (although nonnormality corrections can still be
 ##' requested), no variables in the model nor auxiliary variables specified in
 ##' \code{aux} can be declared as \code{ordered}.
 ##'
 ##' @aliases auxiliary lavaan.auxiliary cfa.auxiliary sem.auxiliary growth.auxiliary
-##' @importFrom lavaan lavInspect parTable
+##' @importFrom lavaan lavInspect parTable lavNames
 ##' @importFrom stats cov quantile
 ##'
 ##' @param model The analysis model can be specified with 1 of 2 objects:
@@ -35,14 +34,21 @@
 ##' @param fun \code{character}. Name of a specific lavaan function used to fit
 ##'   \code{model} to \code{data} (i.e., \code{"lavaan"}, \code{"cfa"},
 ##'   \code{"sem"}, or \code{"growth"}). Only required for \code{auxiliary}.
-##' @param ... additional arguments to pass to \code{\link[lavaan]{lavaan}}.
+##' @param ... Additional arguments to pass to \code{fun=}.
+##' @param envir Passed to \code{\link{do.call}}.
+##' @param return.syntax \code{logical} indicating whether to return a
+##'   \code{character} string of \code{\link[lavaan]{model.syntax}} that can be
+##'   added to a target \code{model=} that is also a \code{character} string.
+##'   This can be advantageous, for example, to use add saturated correlates to
+##'   a \pkg{blavaan} model.
 ##'
 ##' @author
 ##' Terrence D. Jorgensen (University of Amsterdam; \email{TJorgensen314@@gmail.com})
 ##'
-##' @references Enders, C. K. (2008). A note on the use of missing auxiliary
-##'   variables in full information maximum likelihood-based structural equation
-##'   models. \emph{Structural Equation Modeling, 15}(3), 434--448.
+##' @references
+##'   Enders, C. K. (2008). A note on the use of missing auxiliary variables in
+##'   full information maximum likelihood-based structural equation models.
+##'   \emph{Structural Equation Modeling, 15}(3), 434--448.
 ##'   \doi{10.1080/10705510802154307}
 ##'
 ##' @return a fitted \code{\linkS4class{lavaan}} object.  Additional
@@ -87,24 +93,29 @@
 ##' ##       is the independence model plus saturated auxiliary parameters
 ##' lavInspect(fitaux2@external$baseline.model, "free")
 ##'
+##'
 ##' @export
-auxiliary <- function(model, data, aux, fun, ...) {
+auxiliary <- function(model, data, aux, fun, ...,
+                      envir = getNamespace("lavaan"), return.syntax = FALSE) {
   lavArgs <- list(...)
-  lavArgs$data <- data
+  lavArgs$data <- substitute(data)
   lavArgs$fixed.x <- FALSE
-  lavArgs$missing <- "fiml"
-  lavArgs$meanstructure <- TRUE
+  if (fun %in% c("lavaan","cfa","sem","growth")) {
+    lavArgs$missing <- "fiml"
+    lavArgs$meanstructure <- TRUE
+  }
   lavArgs$ordered <- NULL
 
   if (missing(aux))
     stop("Please provide a character vector with names of auxiliary variables")
   if (missing(data))
     stop("Please provide a data.frame that includes modeled and auxiliary variables")
-  if (!all(sapply(data[aux], is.numeric)))
+  if (fun %in% c("lavaan","cfa","sem","growth") && !all(sapply(data[aux], is.numeric)))
     stop("missing = 'FIML' is unavailable for categorical data")
 
 
   PTcols <- c("lhs","op","rhs","user","block","group","free","label","plabel","start")
+  #TODO: add any? if fun %in% c("blavaan","bcfa","bsem","bgrowth")
   ## check parameter table, or create one from syntax
   if (is.list(model)) {
 	  if (any(model$exo == 1))
@@ -118,7 +129,7 @@ auxiliary <- function(model, data, aux, fun, ...) {
 	    startArgs <- lavArgs
 	    startArgs$model <- model
 	    startArgs$do.fit <- FALSE
-	    model$start <- parTable(do.call(fun, startArgs, envir = getNamespace("lavaan")))$start
+	    model$start <- parTable(do.call(fun, startArgs, envir = envir))$start
 	  }
 
 	  missingCols <- setdiff(PTcols, names(model))
@@ -129,8 +140,8 @@ auxiliary <- function(model, data, aux, fun, ...) {
 	} else if (is.character(model)) {
 	  ptArgs <- lavArgs
 	  ptArgs$model <- model
-	  ptArgs$do.fit <- FALSE
-	  PT <- parTable(do.call(fun, ptArgs, envir = getNamespace("lavaan")))[PTcols]
+	  ptArgs$do.fit <- FALSE #FIXME: does this work with blavaan?
+	  PT <- parTable(do.call(fun, ptArgs, envir = envir))[PTcols]
 	} else stop("The 'model' argument must be a character vector of",
 	            " lavaan syntax or a parameter table")
 
@@ -143,13 +154,14 @@ auxiliary <- function(model, data, aux, fun, ...) {
   } else CON <- data.frame(NULL)
 
   ## variable names
-  varnames <- lavaan::lavNames(PT, type = "ov")
+  varnames <- lavNames(PT, type = "ov")
 	if (length(intersect(varnames, aux))) stop('modeled variable declared as auxiliary')
 
 	## specify a saturated model for auxiliaries
 	covstruc <- outer(aux, aux, function(x, y) paste(x, "~~", y))
 	satMod <- c(covstruc[lower.tri(covstruc, diag = TRUE)], paste(aux, "~ 1"), # among auxiliaries
 	            outer(aux, varnames, function(x, y) paste(x, "~~", y)))        # between aux and targets
+	if (return.syntax) return(satMod)
 	satPT <- lavaan::lavaanify(satMod, ngroups = max(PT$group))[c("lhs","op","rhs",
 	                                                              "user","block","group")]
 
@@ -157,7 +169,7 @@ auxiliary <- function(model, data, aux, fun, ...) {
 	mergedPT <- lavaan::lav_partable_merge(PT, satPT, remove.duplicated = TRUE, warn = FALSE)
 	nAuxPar <- nrow(mergedPT) - nrow(PT)
 	newRows <- 1L:nAuxPar + nrow(PT)
-	##FIXME:  mergedPT$user[newRows] <- 2L (list as constraints to omit printing?) or new code (9L)?
+	#TODO:  mergedPT$user[newRows] <- 2L (list as constraints to omit printing?) or new code (9L)?
 	mergedPT$free[newRows] <- 1L:nAuxPar + max(PT$free)
 	mergedPT$plabel[newRows] <- paste0(".p", 1L:nAuxPar + nrow(PT), ".")
 	## calculate sample moments as starting values (recycle over groups)
@@ -186,58 +198,77 @@ auxiliary <- function(model, data, aux, fun, ...) {
 	#   ## match order of parameters in syntax above
 	#   mergedPT$start[newRows] <- startVals
 	# }
-	lavArgs$model <- lavaan::lav_partable_complete(rbind(mergedPT, CON))
-  result <- do.call(fun, lavArgs, envir = getNamespace("lavaan"))
+	lavArgs$model <- lavaan::lav_partable_complete(rbind(mergedPT, CON)) #FIXME for blavaan
+  result <- do.call(fun, lavArgs, envir = envir)
 
-  ## specify, fit, and attach an appropriate independence model
-  baseArgs <- list()
-  baseArgs$model                  <- lavaan::lav_partable_complete(satPT)
-  baseArgs$data                   <- data
-  baseArgs$group                  <- lavArgs$group
-  baseArgs$group.label            <- lavArgs$group.label
-  baseArgs$missing                <- "fiml"
-  baseArgs$cluster                <- lavArgs$cluster
-  baseArgs$sample.cov.rescale     <- lavArgs$sample.cov.rescale
-  baseArgs$estimator              <- lavArgs$estimator
-  baseArgs$information            <- lavArgs$information
-  baseArgs$se                     <- lavArgs$se
-  baseArgs$test                   <- lavArgs$test
-  baseArgs$bootstrap              <- lavArgs$bootstrap
-  baseArgs$control                <- lavArgs$control
-  baseArgs$optim.method           <- lavArgs$optim.method
+  if (fun %in% c("lavaan","cfa","sem","growth")) {
+    ## specify, fit, and attach an appropriate independence model
+    baseArgs <- list()
+    baseArgs$model                  <- lavaan::lav_partable_complete(satPT)
+    baseArgs$data                   <- data
+    baseArgs$group                  <- lavArgs$group
+    baseArgs$group.label            <- lavArgs$group.label
+    baseArgs$missing                <- "fiml"
+    baseArgs$cluster                <- lavArgs$cluster
+    baseArgs$sample.cov.rescale     <- lavArgs$sample.cov.rescale
+    baseArgs$estimator              <- lavArgs$estimator
+    baseArgs$information            <- lavArgs$information
+    baseArgs$se                     <- lavArgs$se
+    baseArgs$test                   <- lavArgs$test
+    baseArgs$bootstrap              <- lavArgs$bootstrap
+    baseArgs$control                <- lavArgs$control
+    baseArgs$optim.method           <- lavArgs$optim.method
 
-  result@external$baseline.model  <- do.call("lavaan", baseArgs, envir = getNamespace("lavaan"))
-  result@external$aux             <- aux
-  result@external$baseline.syntax <- satMod
+    result@external$baseline.model  <- do.call("lavaan", baseArgs, envir = envir)
+    result@external$aux             <- aux
+    result@external$baseline.syntax <- satMod
+  } # else something similar for blavaan?  Better to fit manually for blavFitIndices
+
 	result
 }
 
 ##' @rdname auxiliary
 ##' @aliases lavaan.auxiliary
 ##' @export
-lavaan.auxiliary <- function(model, data, aux, ...) {
-	auxiliary(model = model, data = data, aux = aux, fun = "lavaan", ...)
+lavaan.auxiliary <- function(model, data, aux, ...,
+                             envir = getNamespace("lavaan")) {
+  mc <- match.call(expand.dots = TRUE)
+  mc$fun <- "lavaan"
+  mc[[1L]] <- quote(semTools::auxiliary)
+  eval(mc, parent.frame())
 }
 
 ##' @rdname auxiliary
 ##' @aliases cfa.auxiliary
 ##' @export
-cfa.auxiliary <- function(model, data, aux, ...) {
-	auxiliary(model = model, data = data, aux = aux, fun = "cfa", ...)
+cfa.auxiliary <- function(model, data, aux, ...,
+                          envir = getNamespace("lavaan")) {
+  mc <- match.call(expand.dots = TRUE)
+  mc$fun <- "cfa"
+  mc[[1L]] <- quote(semTools::auxiliary)
+  eval(mc, parent.frame())
 }
 
 ##' @rdname auxiliary
 ##' @aliases sem.auxiliary
 ##' @export
-sem.auxiliary <- function(model, data, aux, ...) {
-	auxiliary(model = model, data = data, aux = aux, fun = "sem", ...)
+sem.auxiliary <- function(model, data, aux, ...,
+                          envir = getNamespace("lavaan")) {
+  mc <- match.call(expand.dots = TRUE)
+  mc$fun <- "sem"
+  mc[[1L]] <- quote(semTools::auxiliary)
+  eval(mc, parent.frame())
 }
 
 ##' @rdname auxiliary
 ##' @aliases growth.auxiliary
 ##' @export
-growth.auxiliary <- function(model, data, aux, ...) {
-	auxiliary(model = model, data = data, aux = aux, fun = "growth", ...)
+growth.auxiliary <- function(model, data, aux, ...,
+                             envir = getNamespace("lavaan")) {
+  mc <- match.call(expand.dots = TRUE)
+  mc$fun <- "growth"
+  mc[[1L]] <- quote(semTools::auxiliary)
+  eval(mc, parent.frame())
 }